Description of the function

Description of the function

mod_lasso

Arguments

response
a vector response variable

regressors
a quantitative matrix of regressor

group
a vector with two levels. (The group of the ANCOVA)

a
the parameters that indicate how much the coefficients will be fused

lambda
if the user wants to use it owns values of lambdas

Format

An object of class R6ClassGenerator of length 24.

Value

The coefficients of the fused lasso ANCOVA for the different value of lambda

Examples

B <- c(1, -1, 1.5, 1.5, rep(0, 6), 2, 0, 2, 0)
group <- c(rep('M1', 10), rep('M2', 10))
regressors <- matrix(rnorm(6*20), ncol = 6)
X  <- model.matrix(~group + group:regressors - 1)
y <- X%*%B + rnorm(20)
y <- scale(y)
mod <- fl2(y, regressors, group)

#> Error in fl2(y, regressors, group): impossible de trouver la fonction "fl2"

colors <- c(rep("grey",2), rep('green',2),rep('black', 6), rep(c("orange","blue"), 2), 'darkgreen', rep('yellow',3), rep('purple',2))
matplot(mod$lambda ,t(mod$beta),type='l',col=colors)

#> Error in as.matrix(x): objet 'mod' introuvable

#' @import R6 Matrix gglasso tidyverse glmnet stabs magrittr viridis stringr FusedLasso